Move examples from Test to examples.
