+++ examples/example_pytorch.py
@@ -1,0 +1,94 @@
+#!/usr/bin/env python3
+"""
+The copyright in this software is being made available under this Software
+Copyright License. This software may be subject to other third party and
+contributor rights, including patent rights, and no such rights are
+granted under this license.
+Copyright (c) 1995 - 2019 Fraunhofer-Gesellschaft zur Förderung der
+angewandten Forschung e.V. (Fraunhofer)
+All rights reserved.
+Redistribution and use in source and binary forms, with or without
+modification, are permitted for purpose of testing the functionalities of
+this software provided that the following conditions are met:
+*     Redistributions of source code must retain the above copyright notice,
+this list of conditions and the following disclaimer.
+*     Redistributions in binary form must reproduce the above copyright
+notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution.
+*     Neither the names of the copyright holders nor the names of its
+contributors may be used to endorse or promote products derived from this
+software without specific prior written permission.
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
+CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
+INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR
+CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+NO EXPRESS OR IMPLIED LICENSES TO ANY PATENT CLAIMS, INCLUDING
+WITHOUT LIMITATION THE PATENTS OF THE COPYRIGHT HOLDERS AND
+CONTRIBUTORS, ARE GRANTED BY THIS SOFTWARE LICENSE. THE
+COPYRIGHT HOLDERS AND CONTRIBUTORS PROVIDE NO WARRANTY OF PATENT
+NON-INFRINGEMENT WITH RESPECT TO THIS SOFTWARE.
+"""
+import torch
+import deepCABAC
+import torchvision.models as models
+import numpy as np
+from tqdm import tqdm
+
+
+def main():
+    # encoding
+    model = models.resnet18(pretrained=True)
+    encoder = deepCABAC.Encoder()
+
+    interv = 0.1
+    stepsize = 2**(-0.5*15)
+    stepsize_other = 2**(-0.5*19)
+    _lambda = 0.
+
+    for name, param in tqdm(model.state_dict().items()):
+        if '.num_batches_tracked' in name:
+            continue
+        param = param.cpu().numpy()
+        if '.weight' in name:
+            encoder.encodeWeightsRD(param, interv, stepsize, _lambda)
+        else:
+            encoder.encodeWeightsRD(param, interv, stepsize_other, _lambda)
+
+    stream = encoder.finish().tobytes()
+    print("Compressed size: {:2f} MB".format(1e-6 * len(stream)))
+    with open('weights.bin', 'wb') as f:
+        f.write(stream)
+
+    # decoding
+    model = models.resnet18(pretrained=False)
+    decoder = deepCABAC.Decoder()
+
+    with open('weights.bin', 'rb') as f:
+        stream = f.read()
+
+    decoder.getStream(np.frombuffer(stream, dtype=np.uint8))
+    state_dict = model.state_dict()
+
+    for name in tqdm(state_dict.keys()):
+        if '.num_batches_tracked' in name:
+            continue
+        param = decoder.decodeWeights()
+        state_dict[name] = torch.tensor(param)
+    decoder.finish()
+
+    model.load_state_dict(state_dict)
+
+    # evaluate(model)
+
+
+if __name__ == '__main__':
+    main()
